Make it easier to identify the Pd flavor using defined symbols in m_pd.h.
This is for the benefit of 3rd party external libraries which may need
to know which Pd flavor they're working with. We've had the PDL2ORK
symbol in m_pd.h for a while, but with DISIS Pd-l2ork 2 that's no longer
unambiguous, so we added the PURR_DATA symbol now.

We also added a new PD_FLAVOR symbol which is defined to be a string
describing the Pd flavor. This is currently only present in Purr Data's
m_pd.h header, of course, but it might conceivably be picked up by other
active flavors such as plugdata in the future.