Fix crash in gui => canvas => append.
The crash was being triggered by last_thing being null when the append
callback was invoked. I noticed this, e.g., with ELSE's else and message
help patches, which otherwise look pretty innocent.

So it seems that last_thing *can* in fact be null when the append
callback is being invoked in certain situations. Whatever the cause may
be, the canvas gui crashing is definitely *not* an acceptable outcome in
*any* situation. This commit deals with that condition by effectively
turning the operation into a no-op, which seems to be the only thing
that we can possibly do in this case.