Make the -legacy flag canvas-local.
[declare -legacy] enables legacy positioning as with the global -legacy
flag, but on a canvas-local basis. If not set in a patch, the flag
defaults to its global counterpart. Note that there is no argument, so
you can only enable this, not disable it. Moreover, once set, the flag
is inherited by all subpatches.

This declaration allows you to use vanilla subpatches in Purr Data where
iemguis are positioned in a slightly different way, which may cause gop
patches created in vanilla to be rendered incorrectly.

Adding [declare -legacy] to the vanilla subpatch gets rid of these
issues, while the rest of your patch is rendered as usual. Note that the
global -legacy startup option will affect *all* patches loaded by Purr
Data, which doesn't help if only some subpatches must be rendered in a
vanilla-like fashion.