Change to nw.js 0.42.0 on (almost) all platforms.
The patch window rendering of the nw.js 0.71.0 version (employed on the
Mac in the last couple of releases) is horribly slow. In fact it's so
utterly slow that it becomes impossible to run animations with the new
pd-lua graphics interface, unless the update frequency is dialed down by
at least an order of magnitude.

Earlier nw.js versions <= 0.42.0 don't have that defect. 0.42.0 also
happens to work reasonably well on macOS, as tested on Big Sur (Intel)
and Ventura (M1). So we downgrade to that version on the Mac for now.

While, we're at it, we also upgrade the Linux and Windows builds to the
same nw.js version, which makes maintaining the JS gui a lot easier.

This might cause some regressions on platforms like rpi, though, where
due to a lack of official binaries we're stuck on older nw.js versions;
but this has already been the case for quite some time, so there's hope
that the regressions at least don't get much worse.