Change to nw.js 0.42.0.
nw.js versions 0.42.4 and later have a serious performance regression in
the JavaScript interpreter. Versions <= 0.42.0 don't have that defect.
0.42.0 seems to work reasonably well on macOS, as well as on Linux and
Windows, so we use that version on all three platforms now, which makes
maintaining the JS GUI a lot easier.

This might cause some regressions on "legacy" platforms like the
Raspberry Pi, though, where due to a lack of official binaries we're
stuck on older nw.js versions; but this has already been the case for
quite some time.