Update nw.js once again and work around the NW2 performance issues.
Turns out that nw.js 0.42.0 crashes a lot on newer macOS versions, so we
need to upgrade to a later version. As Ico figured out a while ago, the
performance issues in later versions are due to the NW2 upgrade in
0.42.4 (https://nwjs.io/blog/nw2-mode/). This can be disabled which gets
JS execution performance back up to where it was in older versions.

Thus we have the following changes:

- Use --disable-features=nw2 to work around the NW2 scripting
  performance regression (https://github.com/nwjs/nw.js/issues/7979).
  This is essentially a backport of Ico's fix in
  https://github.com/pd-l2ork/pd-l2ork/commit/290fffc1. Thanks, Ico!

- Bump nw.js back up to 0.67.1 across platforms (Mac, Linux, Windows).
  This is the same version that Ico uses which seems to work well across
  platforms. (Later versions have some weird window-related regressions
  when used with NW2 disabled, so for now we have to stick to 0.67.1.)