Upgrade nw.js to 0.55.0 (experimental).
Turns out that nw.js 0.42.0 crashes a lot on newer macOS versions, so we
need to upgrade to a later version. As Ico figured out a while ago, the
performance issues in later versions are due to the NW2 upgrade in
0.42.4 (https://nwjs.io/blog/nw2-mode/). This can be disabled which gets
JS execution performance back up to where it was in older versions.

0.55.0 is the last version which doesn't have the nwworkingdir bug on
Linux which causes file dialogs to open in the wrong locations. It also
seems to work just as well (or at least not any worse) than 0.67.x on
Mac and Windows.

The only downside is that 0.55.0 still has the dark mode bug on Linux
(https://github.com/agraef/purr-data/issues/42); Mac and Windows seem to
be fine. But this is a cosmetic issue which can be worked around by
adjusting the desktop theme, while the nwworkingdir bug breaks important
functionality and has no known work-around. So I think that fixing the
latter at the cost of breaking dark mode is the most sensible choice.

Note that if you absolutely need to have dark mode working on Linux, you
can now just use the new patch-nwjs.sh script to patch your Purr Data
installation to use a later version such as 0.67.0 which doesn't have
the dark mode bug.