Fix a bug with connect_as_client_to_secondary_instance and IPv6 addresses.
This is a regression introduced with the nw.js update on Linux and
Windows (Mac seems to be fine), presumably because nw.js prefers IPv6
now, while the C engine still "talks" IPv4. The symptom is that if the
user tries to open a patch through a secondary purr-data instance,
opening the socket on "localhost" in the primary instance fails with a
"connection refused" error.

We now work around this by first calling dns.setDefaultResultOrder
to ensure that IPv4 addresses are preferred in any case.

In addition, we now print an error message if the connection attempt in
the primary instance fails for some other reason, so that the user at
least knows that something went wrong. The error message will be
somewhat cryptic, but decipherable by any network expert. :)