Downgrade nw.js to 0.42.0 on Linux.
This gets rid of some regressions with 0.67.1 on this platform.
Specifically, nwworkingdir is working in 0.42.0, but is pretty much
broken in 0.67.1 on Linux (works on Mac and Windows, though). This
makes the Open File dialog open in random places rather than the
directory set by nwworkingdir, which is particularly annyoing in the
help browser, where we rely on nwworkingdir to set the right directory
in the documentation folder.

As we're not using NW2 anyway, 0.67.1 doesn't really offer us any
compelling advantages on Linux, as far as I can tell. So going back to
0.42.0 which works well on Linux, while using 0.67.1 on the other
platforms, seems to be a reasonable compromise to get a good user
experience across all platforms.