Fix the cords in graph_vis.
The corresponding vanilla code doesn't seem to need this, apparently the
cords are adjusted as soon as the gop rectangle is redrawn in the
parent. But in purr-data that doesn't seem to happen, so you'll get
"dangling cords" if the size of a gop area changes, and these stick
around on the canvas even after redrawing the gop area.

I've seen this many times while creating a new gop subpatch, especially
since that's become so easy to do now with Guillem's Encapsulate menu
operation. The problem goes away as soon as you move the subpatch window
around on the canvas, but that shouldn't be needed.

So we now call canvas_fixlinesfor() to have the cords fixed as soon as
the gop area is redrawn in the parent. Problem solved.