DRAFT: fake listbox object.
This is in fact just a symbol gatom masquerading as a vanilla listbox,
so that you can open vanilla patches containing such objects. It doesn't
offer all the bells and whistles in the GUI interface of vanilla's
listbox -- editing is the same as with the symbol gatom. Which is
workable, but nothing to write home about. Other than that, it provides
the same methods as the vanilla listbox, and looks the part.