Implemented dragging numbers in a listbox.
This works pretty much like in vanilla, but you can change a number in a
listbox by either clicking on a number and then dragging upward or
downward, or by using the cursor up and down keys and then pressing the
Return key. In either case, pressing the Shift key changes the numbers
in 0.01 increments. Thus it works pretty much the same as with Purr
Data's number gatom.