Got rid of gatom's "#keyname_a" binding which isn't needed anymore.
It also interfered with listbox's list input, which expects a proper
list value instead of a key event. So this is resolved now.

I don't see why this was ever needed after introducing grabx() to
install a proper key handler when clicking on a gatom. Maybe it was a
left-over from the time before that got added. And even then it caused
issues (dangling "#keyname_a" binding when a gatom got deleted causing
crashes; spurious "#keyname_a: couldn't unbind" errors in the
console). These are all gone now.

Also gave the iemgui numbox the same treatment. While it didn't have the
same regressions as gatom, the "#keyname_a" receiver isn't needed there
either.