Additional fixes to the Windows installer.
During an unattended install (/SILENT /SUPPRESSMSGBOXES), the uninstall
file keeps changing names (unins000.exe, unins001.exe, etc.), which we
now accommodate for. Otherwise the installer would not find the
uninstall executable even if it's there.

In addition, to improve usability of the installer, we now bail out if
an existing installation is found, but running the uninstaller fails for
some reason, which means we have no way of removing the existing
installation in an automatic fashion. Just manually removing the
installation folder is then much quicker than clicking through the
endless barrage of messages and prompts you're subjected to when the
installer just continues anyway.

NB: The installer still only checks for the unins00?.exe and pd.exe
files to determine whether there is an existing installation, which is a
bit simplistic, and thus the user can fool the installer into thinking
that there is no existing installation when most of the files are
actually there. But then he gets what he asked for. There are better
ways to go about this uninstall business, but I don't really know enough
about innosetup to make them work, so we'll keep it simple for now.