Fix escaping of comma and semicolon in FUDI conversion.
Fixes https://github.com/agraef/purr-data/issues/24

The issue was that lookahead instead of lookback was used in the regex
for detecting preceding backslashes. Thus comma and semicolon would
always be substituted when editing an object or message on the canvas,
even when preceded by a backslash, since the lookahead was trivially
satisfied.