Revert bogus change in binbuf_error() from rev. f3d1905.
This was a misguided attempt to fix a simple warning complaining about a
string buffer being passed directly into vsnprintf() (which is easy to
avoid, just pass the buffer as a %s argument instead). The issue was that
it instead attempted to pass the va_list directly to another variadic
function, which you can't do.

This resulted in garbled error messages from binbuf_eval() in case of
any error conditions, which are now fixed.