Some cosmetic changes to the audio and midi prefs.
This adds vanilla-like 44.1k, 48k, and 96k buttons for the sample rate
input, which are very convenient.

To make this look nice, we also make sure that the text inputs and
selects at the top of the audio prefs all properly line up. The problem
there is that even at the same width, the select widgets are always
shorter than the inputs, which has bothered me for a long time. To deal
with this, we now set the width in px on these elements explicitly.

Note that this overrides the width of the rate input set in the css
files, but that setting seems inconsequential anyway, so we happily
override it. (We might want to remove that css width setting in the
future.)

Also, we don't restrict the width of the audio api select, so that this
can grow depending on the platform-specific options.

We give a similar treatment to the midi prefs which look all topsy-turvy
by default for the ALSA API. To fix this, we make the two span elements
at the top of the midi prefs the same width.

Finally, while we're at it, we also remove the Pd-l2ork branding in the
title bar of the prefs dialog.