Fix invalid audio and midi device indices in the preferences.
This may happen when changing audio and midi apis on the fly. We now
detect and fix such issues when the prefs dialog is populated.

The main culprit here seems to be the OSS MIDI backend on Linux. Now you
might argue that it's not worth bothering with OSS on Linux these days
anymore, and you'd be right. But as long as this backend is still on
tap, the user shouldn't see any garbage device entries in the prefs just
because the driver acts up and/or the engine passes wrong device numbers
to the gui in certain situations. This should now be fixed.