Partially revert "remaining things".
This partially reverts commit 5a38eee77230f7f974ad60c5bcfb34bf4f2c10fb.

The commit causes the wiggly lines to appear (indicating no existing
completions) as soon as you start entering arguments, which is not what
we want. Also, the lower extension of the object box to make room for
the wiggly lines looks a bit out of whack.

The chunk that we keep from that commit is L1035..1041 where the
scrollbar of the popup menu gets updated.