Allow subdirectories in the directory view of the help browser.
These will be indicated by a trailing slash and switch the directory in
the help browser when clicked, which facilitates navigating the help
directory tree. It also makes it possible to have directories which just
contain other subdirectories in the toc or a bookmark (previously these
would just come up empty when clicked).