Backport rev. 05178840 from vanilla (glob_open).
See https://github.com/pure-data/pure-data/commit/05178840

This adds an extra flag to the global "open" message indicating whether
a patch should just be made visible, or be loaded twice if it is already
open. The latter is used by most open operations, the former for help
patches opened from the browser or the menu, including the testtone and
load-meter patches. (This also fixes an issue with the testtone patch
which would spit out a bunch of error messages about multiply defined
symbols if it was accidentally opened more than once.)