Improve the handling of various file types in pdgui.doc_open.
Previously we threw anything we didn't recognize at the engine to try
and open it as a patch, which wasn't a good idea. Now we only pass .pd
files to the engine, and anything else we don't recognize we leave up to
Shell.openItem(), which should hopefully know how to deal with it.

TODO: In the future we might want to use file(1) magic patterns to
make a better guess at whether a given file is in fact a Pd patch.