Help browser: Add some mimetypes to the file picker.
Now that we only pass proper .pd files to the engine, there is no good
reason that we still need to be as restrictive with the file picker as
before. The help browser intentionally focuses on the most important
items, basically just patches and text docs. But the file picker should
allow the user to open other related items, such as audio and image
files, which might accompany the patches, while still filtering out
irrelevant file types like shared libraries that one might also find in
external directories.

There are a lot more mimetypes that we could add here. For starters, we
listed a few that the user might come across in the docs or in a typical
Pd project.