Fix a bug with editing pddplink and helplink objects (#48).
Also fix the incoming cord to helplink's hidden inlet when deactivating
or displacing the object. Otherwise the cord would just disappear or
fail to move along, respectively.

NB: Apparently, the cords are only handled automagically for regular
text objects inside a box which helplink doesn't have. pddplink isn't
affected by this, because pddplink -box is a regular object, and the
pddplink without the box doesn't have an inlet.