Reduce flickering in the xlet hover animation.
This animation actually looks quite nice, at least in the default.css
theme, but it has a serious defect which causes a lot of flickering in
certain situations -- that is, if you have an array in your patch AND
dsp is enabled AND the graph is constantly being updated (using metro +
tabwrite~) AND the object with the outlet being hovered over is
selected.

When the flickering occurs, it is at best very distracting, and may
at worst cause harm to people who are sensitive to this.

I don't know why this happens, so I adjusted the parameters of the
webkit animation to mitigate the flickering to a point where I don't
notice it anymore. This is implemented in both the default.css and
inverted.css themes for now.

As I'm too lazy to adjust the animations in all the other css files to
follow suit, I just disabled those for now. I don't think that this is a
big loss, since there are other visual cues which make it clear that
you're over an xlet which may serve as source or destination of a patch
cord, such as changes of the mouse cursor or the xlet changing color
when hovering over it. But if anyone wants to take the time and fix all
these then please be my guest.