More updates and typo fixes in the all_about_expr_functions help patch.
Re-added the avg and Avg array functions, which were removed in an
earlier version, but were added back in expr 0.56.

Note that the acoustic and string functions are new in expr 0.57 and
0.58, these still need to be documented in the help patch. For the time
being, we link to the latest html documentation in ../8.topics/expr.htm
from vanilla Pd, which is up-to-date.