Add the 'Meet the Cat' tutorial to the distribution.
The about dialog has offered a link to the online version of the
tutorial for a long time, but we've considered to also include an
off-line version of this document for easier access.

However, I wanted to include this as a single, self-contained file (pdf,
that is), and the pandoc-formatted pdf never looked all that great. But
since we now switched to Typora instead, that problem is solved, so we
added this now and have it linked as a local file in the about patch.