pdgui.js: Rework file_dialog() so that it uses get_pd_opendir().
This now uses the same default location as the file dialogs in the menu.
As far as I can tell, this matches vanilla's behavior and in any case it
seems the most sensible choice.

We also check that the specified path actually exists and is a
directory. This isn't guaranteed if file_dialog() gets invoked
programmatically, through openpanel/savepanel. It's better to fall back
to get_pd_opendir() in this case, rather than being sent to some random
location that the system thinks is a good default (it usually isn't).