portmidi(macosx): Backport some backend fixes from the latest source.
These are needed to make the CoreMIDI device refresh work on macOS.

This isn't a full upgrade. It just backports the required changes in
pm_macosxcm_init to our source. A full upgrade of our bundled portaudio
and portmidi sources is probably in order, but this can wait until
vanilla Pd leads the way.