Fix some issues with gatom and iemgui editing and shift status.
Shift status wasn't really working correctly in gatoms and iemgui
numboxes, since Shift key up events were never reported to the
keynameafn function, so that the shift status recorded in the object,
once turned on, never changed.

This is now fixed, so releasing the shift key while dragging will switch
back to unit increments. The only downside is that you now have to keep
pressing the Shift key during a mouse drag as long as you want to change
the value in 100ths increments. Buth this is also how the corresponding
arrow key up and down commands have always worked, so I think that this
behavior is correct and what users would expect.

Moreoever, symbol and listbox atoms now always populate the edit buffer
with the current value when shift-clicked; previously, the buffer
wouldn't be filled after the initial shift-click, so you'd have to
retype the entire value first, which was a big nuisance.

Also, thanks to the fixes in shift status reporting described above, the
backspace key now also works correctly. That is, it deletes the last
character when unshifted, and the entire value when shifted; previously
it would just always erase the entire value once the Shift key had been
pressed, making it basically impossible to edit the value after pressing
Shift just once.

Last but not least, I also fixed some typos and bad indentation, and
removed some commented debug statements.