Fix another issue with line breaks in canvas objects.
Our browser engine does seem to like line breaks, so it adds lots of
them were they weren't called for. This isn't an issue with most kinds
of objects since the binbuf parser will eat those extra newlines, but
with visual formatting while editing comments it is.

So we remove the extra newlines now. The catch here is that there's no
way to tell whether a line break was deliberately added by the user or
inserted automatically by the browser.

This means that two or more consecutive empty lines will always be
reduced to just a single empty line. However, there's a work-around:
Add a single blank to an empty line, and it will always be preserved. I
think that for the time being this is a workable solution.