Update the Pd html manual to the version from Pd 0.53-2.
Our version of the manual was getting a bit long in the tooth, so an
update was in order.

I also backported our pd-l2ork-specific additions to the manual, and
fiddled a bit with the css to make it look nicer.

Other than that, it's just the manual from 0.53-2, which was the last
version before the addition of multi-channel signals, which we don't
have yet. This should be as close to our engine as it gets, so that most
of the technical information should be applicable.