Fix up end-of-line whitespace regexes.
This was an ill-fated attempt to employ variable-length lookback in
order to make the regex "count" the parity of consecutive backslashes
preceding trailing whitespace, like the Pd binbuf parser does. But that
doesn't work with the JS RE engine at least, so I removed those bits
again.

What this means is that trailing whitespace will be kept in some cases
where the binbuf parser might have removed it, which shouldn't have any
grave consequences (fingers crossed). The important thing here is that
escaped trailing whitespace is kept and does not suddenly disappear when
editing an object, and that still works with the simplified regexes.