Fix intelligent patching mode 4.
This is the multi-connect mode where you select 3 or more objects,
connect any two of them, and then hope for the best. ;-)

This never worked for me. You know that there's something wrong if a
feature needs more than a few lines of text to explain it. The outcomes
were often unpredictable, and I always considered the visual order being
used (left-to-right then top-to-bottom) to be the wrong one.

I corrected this now, so that a top-to-bottom then left-to-right order
of the selection is used, and the choice of connections is more limited:
if y1--y2 is the initial connection, then only sources to the right of
y1 and targets to the right of y2 will be considered, and the operation
will just fail if y2 comes before y1 in the visual order. This also
makes it much easier to paint a mental image of the operation in terms
of two rows of objects, the source row containing y1 above, and the
target row containing y2 below it.

I think that this works in a more intuitive fashion, and while the
possible connections are more limited, they are also much more
predictable. The operation still does either fan-out or fan-in, based
on what will give you the most connections, breaking ties in favor of
fan-outs. It also still reverses the choice if you hold ctrl while
connecting, but it seems that the operation will now do the right thing
most of the time without that crutch.