Improvements to intelligent patching mode 4.
Fixed a few more bugs in the multi-connect fan-out and fan-in modes.

Also added a third option to do parallel connections, which is activated
by holding the ctrl key while doing the initial connection. (This was
previously used to reverse the fan-out / fan-in choice, so this
functionality is not available anymore. But it was largely obsolete with
the new algorithm anyway, while the new parallel connection feature
provides genuinely useful new functionality.)

Parallel connections quite simply connect as many selected object pairs
as possible, going left to right starting at the initial y1-y2 pair, and
using the same outlet and inlet numbers as in the initial connection:

y1 y1' y1'' ...
|  |   |
y2 y2' y2'' ...

I have always missed this feature. It's very basic, but useful in so
many situations. While this isn't a multi-connect operation in the same
sense as the fan-out and fan-in operations in mode 4, it uses the same
basic interface and visual top-to-bottom and left-to-right selection
order, so it was most convenient to add it there.