Further improvements to intelligent patching mode 4.
Made the algorithm a lot more flexible by applying some heuristics to
estimate whether the selected objects are primarily layed out in a
vertical (row-based) or in a horizontal (column-based) configuration.

Thus, in addition to the common case of vertical connections, the
algorithm now also deals with horizontal arrangements which occur in
some use cases such as ladder structures.

Furthermore, the algorithm tries to cope with situations where the
target object comes *before* the source object, by reversing the order
of the primary coordinate. E.g., this will allow you to make back
connections running from bottom to top instead of the usual
downward-pointing connections.

This means that mode 4 offers more possibilities for connections now,
which comes at the cost of permitting some corner cases where the
results are less predictable. But the results should still be a lot more
predictable than with the old implementation.