Fix moonlib/dinlet~.c for multi-channel compatibility.
Updated the source once more to the latest version of the t_vinlet data
structure and initialization code from g_io.c which changed with the
multi-channel backport.

NOTE: This object lacks the message-forwarding of the new inlet~ object
introduced in Pd 0.51. It's also a maintenance burden because of the
amount of code it pulls from g_io.c. Thus the code either needs to be
refactored to facilitate using the required parts from the core
(including message forwarding), or this object should be considered for
removal from moonlib in the future.