Fix crash during creation of fiddle~.
Made the fiddle~ message callbacks and sigfiddle_new static, otherwise
instantiation would crash because of a bogus NULL class pointer
(although sigfiddle_class is surely initialized, otherwise sigfiddle_new
could never be called in the first place).

The other stuff in pd/extra seems to be fine, all methods there are
static already, but I haven't checked all the externals.

Observed on Arch with gcc 14.2.1 20240910, Mac and Windows builds seem
to be fine. To me this looks like a compiler or linker bug for that
specific gcc version.  (Come on, Arch, are we now getting compiler
versions from git that haven't been released yet and presumably still
need testing? Who thought that this would be a good idea?)