Backport Alexandre Porres' improved testtone patch from vanilla Pd.
This replaces the old testtone patch from the days of yore with
Alexandre's new version which looks less boring and also offers some
useful new functionality.

Fixed some typos and minor grammatical errors in the comments, while
also reducing the wall of text at the top, so that the patch (which is
substantially larger than the old one) requires a little less precious
screen real-estate. I think that removing some of the audio/MIDI
configuration hints there doesn't do much harm since that information is
also in the instructions subpatch, as well as being discussed at length
in the Pd manual and the "Meet the Cat" tutorial.