Simplified transform matrix in gui_scalar_new.
This uses the transform matrix as is, as Jonathan originally implemented
it, reverting the extra padding Ico added for the different plot styles
in rev. e2e65133 to accommodate the custom scrollbars. Those changes
make it hard to debug array rendering issues, that's why I remove them
for now, but we might re-introduce them again later if needed.