Rework the rich text tag syntax.
Color tags must now start with '=', e.g.: <=red>. The only other tags
recognized are b, h, i, s, and u, optionally preceded by '/' to indicate
an end tag. Escaping tags with an initial '!' works as before. Anything
else enclosed in <...> is taken as a literal.

Note that we want to be as specific as possible here, since help patches
also use <...> in comments as ad-hoc syntax for certain meta variables,
in which case we want them to be treated as literals. The revised tag
syntax should hopefully leave these alone now.